refactor(api): Move `api/db` to `api/core/db`
