+++ .vscode/api.code-snippets
@@ -4,8 +4,7 @@
     "prefix": "query-list",
     "body": [
       "import { GraphQLFieldConfig, GraphQLList, GraphQLNonNull } from \"graphql\";",
-      "import { type Context } from \"../context\";",
-      "import db, { ${1/^(.)(.*?)s?$/${1:/upcase}$2/} } from \"../db\";",
+      "import { Context, db, ${1/^(.)(.*?)s?$/${1:/upcase}$2/} } from \"../core\";",
       "import { ${1/^(.)(.*?)s?$/${1:/upcase}$2/}Type } from \"../types\";",
       "",
       "export const ${1:field}: GraphQLFieldConfig<unknown, Context> = {",
