+++ api/auth/facebook.ts
@@ -4,7 +4,7 @@
 import { RequestHandler } from "express";
 import got from "got";
 import { AuthorizationCode } from "simple-oauth2";
-import { IdentityProvider } from "../db";
+import { IdentityProvider } from "../core";
 import env from "../env";
 import authorize from "./authorize";
 import { createState, verifyState } from "./state";
