+++ api/core/index.ts
@@ -1,5 +1,9 @@
 /* SPDX-FileCopyrightText: 2016-present Kriasoft <hello@kriasoft.com> */
 /* SPDX-License-Identifier: MIT */
 
+export * from "../../db/types"; // Generated db types
+export * from "../context";
+export * from "./db";
+export * from "./handleError";
 export * from "./logging";
 export * from "./storage";
