+++ api/mutations/auth.ts
@@ -8,8 +8,7 @@ import {
   GraphQLObjectType,
   GraphQLString,
 } from "graphql";
-import { Context } from "../context";
-import db, { User } from "../db";
+import { Context, db, User } from "../core";
 import { UserType } from "../types";
 import { validate, ValidationError } from "../utils";
