+++ api/queries/me.ts
@@ -1,9 +1,8 @@
 /* SPDX-FileCopyrightText: 2016-present Kriasoft <hello@kriasoft.com> */
 /* SPDX-License-Identifier: MIT */
 
-import type { GraphQLFieldConfig } from "graphql";
-import type { Context } from "../context";
-import type { User } from "../db";
+import { GraphQLFieldConfig } from "graphql";
+import { Context, User } from "../core";
 import { UserType } from "../types";
 
 /**
