+++ db/knexfile.ts
@@ -3,7 +3,7 @@
 
 import envars from "envars";
 import minimist from "minimist";
-import config from "../api/db/config";
+import config from "../api/core/dbConfig";
 
 // Load environment variables (PGHOST, PGUSER, etc.)
 envars.config({ env: minimist(process.argv.slice(2)).env });
